feat: backend for authentication
